requirements.txt
auto_mix_prep==0.2.0
hydra-core==1.3.2
matplotlib==3.7.1
numpy==1.23.5
PyYAML==6.0
PyYAML==6.0.1
scikit_learn==1.2.2
seaborn==0.12.2
torch==2.0.0
torchvision==0.15.0
wandb==0.14.0
